+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Enable/disable correction interrupt
+- Enable/disable/clear second interrupt
 
 ## [0.2.0] - 2025-08-23